Use full session list in search filters view
